backend/requirements.txt
# ─── Core Django ──────────────────────────────────────────────────────────────
Django==5.2.12
djangorestframework
django-cors-headers
django-allauth
dj-rest-auth
djangorestframework-simplejwt
whitenoise==6.12.0
gunicorn==25.1.0
uvicorn==0.34.3

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.11
pgvector==0.4.2
redis==7.3.0
django-redis

# ─── Authentication & Security ────────────────────────────────────────────────
PyJWT==2.11.0
cryptography==46.0.5
django-axes

# ─── Task Queue ───────────────────────────────────────────────────────────────
celery==5.6.2
kombu==5.6.2
billiard==4.2.4
django-celery-beat
django-celery-results==2.5.1

# ─── AI/ML Services ───────────────────────────────────────────────────────────
openai==2.26.0
tiktoken==0.12.0
tenacity==9.1.4

# ─── HTTP and Networking ──────────────────────────────────────────────────────
requests==2.32.5
httpx==0.28.1

# ─── Environment and Configuration ───────────────────────────────────────────
python-dotenv==1.2.2

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dateutil==2.9.0.post0
PyYAML==6.0.3
tqdm==4.67.3

# ─── Development & Testing ────────────────────────────────────────────────────
colorama==0.4.6

# ─── REMOVED (unused in production) ──────────────────────────────────────────
# numpy, pandas, scikit-learn  — not used anywhere in codebase
# pydantic, pydantic-core      — not used anywhere in codebase
# rich                         — not used anywhere in codebase
# urllib3                      — pulled in automatically by requests

